Update boundary_conditions.py
This redoes the code from pull request #1041 to allow for PyTorch backend only. It raises error on any other backend being used and trying to set up multiple components. Also included is code that was used that allows this to work in other backends, that code it just slow in tensorflow backends. This is added so that when we revisit this problem we have what was tested already.